src/evaluation/virality_evaluator.py
"""Virality evaluation using Firecrawl to scrape forums."""
import sys
import time
import re
from pathlib import Path
from typing import Dict, Any, List, Optional
from firecrawl import FirecrawlApp
import asyncio

# Add parent directory to path
sys.path.append(str(Path(__file__).parent.parent.parent))
from src.config import config
from src.observability.metrics import MetricsLogger

class ViralityEvaluator:
    """Evaluates virality by scraping video forums with Firecrawl."""
    
    def __init__(self, use_wandb: bool = True):
        """Initialize virality evaluator."""
        api_key = config.get_api_key('firecrawl')
        if not api_key:
            print("Warning: FIRECRAWL_API_KEY not found - using simulated virality for demo")
            self.app = None
            self.simulation_mode = True
        else:
            try:
                self.app = FirecrawlApp(api_key=api_key)
                self.simulation_mode = False
            except Exception as e:
                print(f"Warning: Firecrawl initialization failed: {e} - using simulation mode")
                self.app = None
                self.simulation_mode = True
        self.targets = config.get('firecrawl.targets', [])
        self.max_pages = config.get('firecrawl.max_pages', 10)
        self.delay = config.get('firecrawl.delay', 2)
        self.use_wandb = use_wandb
        self.metrics = MetricsLogger() if use_wandb else None
        
        # Cache for scraped data
        self.scraped_data_cache: Dict[str, Any] = {}
    
    def scrape_forums(self, keywords: Optional[List[str]] = None) -> Dict[str, Any]:
        """
        Scrape video virality forums for engagement data.
        
        Args:
            keywords: Optional keywords to focus on
            
        Returns:
            Dictionary with scraped data and virality metrics
        """
        try:
            all_data = []
            
            for target_url in self.targets:
                print(f"Scraping {target_url}...")
                
                try:
                    if self.simulation_mode or not self.app:
                        # Simulate virality data for demo
                        import random
                        # Simulate engagement based on keywords
                        base_score = random.uniform(0.3, 0.8)
                        if keywords:
                            # Boost score if keywords match video content
                            keyword_bonus = len(keywords) * 0.1
                            base_score = min(0.9, base_score + keyword_bonus)
                        
                        print(f"  (Using simulated virality data for demo)")
                        scrape_result = {
                            'markdown': f"Simulated content from {target_url} for demo",
                            'content': f"Simulated engagement data"
                        }
                        metrics = {
                            'engagement_score': base_score * 100,
                            'keywords_matched': keywords or [],
                            'upvotes': int(base_score * 1000),
                            'comments': int(base_score * 100),
                            'mentions': len(keywords) if keywords else 0
                        }
                    else:
                        # Try different API methods based on Firecrawl SDK version
                        try:
                            # Method 1: Direct scrape_url
                            scrape_result = self.app.scrape_url(url=target_url)
                        except (AttributeError, TypeError):
                            try:
                                # Method 2: Using crawl if available
                                scrape_result = self.app.crawl_url(target_url)
                            except:
                                # Fallback - simulate
                                raise Exception("Firecrawl API methods not available")
                        
                        # Extract engagement metrics from scraped content
                        metrics = self._extract_engagement_metrics(scrape_result, keywords)
                    
                    # Handle different response formats
                    if isinstance(scrape_result, dict):
                        scraped_content = scrape_result
                    else:
                        # If result is already the content
                        scraped_content = {'markdown': str(scrape_result), 'content': str(scrape_result)}
                    
                    all_data.append({
                        'url': target_url,
                        'data': scrape_result if isinstance(scrape_result, dict) else {'content': str(scrape_result)},
                        'metrics': metrics,
                        'timestamp': time.time()
                    })
                    
                    # Log metrics
                    if self.metrics:
                        self.metrics.log_scraping_metrics({
                            'url': target_url,
                            'pages_scraped': 1,
                            'engagement_score': metrics.get('engagement_score', 0),
                            'keywords_found': len(metrics.get('keywords_matched', []))
                        })
                    
                    # Delay between requests
                    time.sleep(self.delay)
                
                except Exception as e:
                    print(f"Error scraping {target_url}: {e}")
                    if self.metrics:
                        self.metrics.log_error('scraping', f"Failed to scrape {target_url}: {str(e)}")
                    continue
            
            # Aggregate virality scores
            virality_score = self._calculate_virality_score(all_data)
            
            return {
                'scraped_data': all_data,
                'virality_score': virality_score,
                'total_urls_scraped': len(all_data),
                'success': True
            }
        
        except Exception as e:
            error_msg = f"Scraping failed: {str(e)}"
            
            if self.metrics:
                self.metrics.log_error('virality_evaluation', error_msg)
            
            return {
                'scraped_data': [],
                'virality_score': 0.0,
                'total_urls_scraped': 0,
                'success': False,
                'error': error_msg
            }
    
    def _extract_engagement_metrics(
        self, 
        scraped_content: Dict[str, Any], 
        keywords: Optional[List[str]] = None
    ) -> Dict[str, Any]:
        """
        Extract engagement metrics from scraped content.
        
        Looks for upvotes, comments, likes, shares, etc.
        """
        content = scraped_content.get('markdown', '') or scraped_content.get('content', '')
        
        metrics = {
            'engagement_score': 0.0,
            'keywords_matched': [],
            'upvotes': 0,
            'comments': 0,
            'mentions': 0
        }
        
        if not content:
            return metrics
        
        # Extract numbers (upvotes, comments, etc.)
        # Pattern for Reddit-style: "123 upvotes", "456 comments"
        upvote_pattern = r'(\d+)\s*(?:upvote|upvotes|karma|points)'
        comment_pattern = r'(\d+)\s*(?:comment|comments)'
        
        upvote_matches = re.findall(upvote_pattern, content.lower())
        comment_matches = re.findall(comment_pattern, content.lower())
        
        if upvote_matches:
            metrics['upvotes'] = max([int(m) for m in upvote_matches if m.isdigit()], default=0)
        
        if comment_matches:
            metrics['comments'] = max([int(m) for m in comment_matches if m.isdigit()], default=0)
        
        # Match keywords if provided
        if keywords:
            content_lower = content.lower()
            matched_keywords = [kw for kw in keywords if kw.lower() in content_lower]
            metrics['keywords_matched'] = matched_keywords
            metrics['mentions'] = len(matched_keywords)
        
        # Calculate engagement score
        # Weight: upvotes (0.5), comments (0.3), keyword mentions (0.2)
        metrics['engagement_score'] = (
            metrics['upvotes'] * 0.5 +
            metrics['comments'] * 0.3 +
            metrics['mentions'] * 0.2
        )
        
        return metrics
    
    def _calculate_virality_score(self, all_data: List[Dict[str, Any]]) -> float:
        """Calculate aggregate virality score from all scraped data."""
        if not all_data:
            return 0.0
        
        total_engagement = sum(
            data.get('metrics', {}).get('engagement_score', 0.0)
            for data in all_data
        )
        
        # Normalize to 0-1 scale (adjust threshold as needed)
        normalized_score = min(total_engagement / (len(all_data) * 100), 1.0)
        
        return normalized_score
    
    def evaluate_prompt_virality(self, prompt: str, generated_content: Dict[str, Any]) -> float:
        """
        Evaluate how viral a prompt might be based on scraped data.
        
        Args:
            prompt: The prompt used for generation
            generated_content: Information about generated content
            
        Returns:
            Virality score (0-1)
        """
        # Extract keywords from prompt
        keywords = self._extract_keywords(prompt)
        
        # Scrape with keywords (if available)
        scraping_result = self.scrape_forums(keywords)
        
        # Use virality score
        virality_score = scraping_result.get('virality_score', 0.0)
        
        # If no virality score (all zeros or simulation mode), generate simulated score based on prompt quality
        if virality_score == 0.0 or (hasattr(self, 'simulation_mode') and self.simulation_mode):
            import random
            # Simulate virality based on prompt characteristics
            # Longer, more detailed prompts tend to score higher
            prompt_length_factor = min(len(prompt) / 200, 1.0)  # Normalize by 200 chars
            keyword_factor = len(keywords) * 0.15
            base_score = random.uniform(0.45, 0.75) * prompt_length_factor
            virality_score = min(0.9, base_score + keyword_factor + random.uniform(-0.15, 0.15))
            print(f"  Simulated virality score: {virality_score:.3f} (based on prompt characteristics)")
        
        # Log evaluation
        if self.metrics:
            self.metrics.log_virality_evaluation({
                'prompt': prompt[:100],  # Truncate for logging
                'virality_score': virality_score,
                'keywords_used': keywords
            })
        
        return virality_score
    
    def _extract_keywords(self, text: str) -> List[str]:
        """Extract potential keywords from text."""
        # Simple keyword extraction (can be improved with NLP)
        words = text.lower().split()
        
        # Remove common stop words
        stop_words = {'the', 'a', 'an', 'and', 'or', 'but', 'in', 'on', 'at', 'to', 'for', 'of', 'with', 'is', 'are', 'was', 'were'}
        keywords = [w for w in words if w not in stop_words and len(w) > 3]
        
        # Return top 5 keywords
        return list(set(keywords))[:5]

