requirements.txt
# Core dependencies
openai>=1.12.0
anthropic>=0.18.0
openai-whisper>=20231117
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
diffusers>=0.25.0
accelerate>=0.25.0

# APIs and Services
firecrawl-py>=0.0.16
wandb>=0.16.0
composio-core>=0.3.0
replicate>=0.25.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0
numpy>=1.24.0
pillow>=10.1.0
aiohttp>=3.9.0
psutil>=5.9.0

# Audio processing
ffmpeg-python>=0.2.0
pydub>=0.25.1

# RL
gym>=0.26.0
stable-baselines3>=2.2.0

